pos_payment_terminal: receive transaction refs
Provide a mechanism to send the order UID to the payment terminal,
which can then pass the transaction reference generated by the
payment provider back to Odoo, and which is then added to the
order payment lines. This allows their subsequent reconciliation.
The order is also automatically validated when the payment
finishes.