[IMP] pos_cancel_reason: Canceled orders with zero value
There was a problem when canceling orders: they were going to the
backend with the full amount - despite the canceled status - and it
ended up confusing the sum of sales for a session, since they were
considered.

This implementation causes orders that were canceled to run out of order
lines, and their items become the model of the canceled line. That wa
there is no problem handling these requests in the backend.