[IMP] pos_event_sale: allow to negate lines
Negated lines are particularly useful with PoS modules that disable the removal
of pos orderlines through the use of ``disallowLineQuantityChange``.

This is the case of `l10n_fr_pos_cert` module, for example. But there are other
core modules doing it, too.

In this case, added order lines can't be removed. Instead, the cashier is
encouraged to add a negative to 'cancel' the original one.

e.g.:

Adds a line with 10 tickets.
Then, adds a line with -10 tickets.

This should be almost the same as removing the first line: no actual
registration should be confirmed in the end, as we're reverting this sale.