[FIX] pos_order_return: empty returns
If by mistake a cashier creates an empty return, the order won't be
synched as the backend code will raise an exception. The order will be
stuck and everytime it tries to sync it will raise an error.

For the moment, avoid trying to create an empty picking return, although
we should forbid it in UX in the first place.

TT46400